requirements.txt
streamlit==1.26.0
requests==2.31.0
fpdf2==2.7.4